+++ MSAL/src/public/MSALAuthenticationSchemePop.h
@@ -27,6 +27,8 @@
 
 #import "MSALAuthenticationSchemeProtocol.h"
 
+@class MSALExternalKeyPair;
+
 NS_ASSUME_NONNULL_BEGIN
 
 @interface MSALAuthenticationSchemePop : NSObject<MSALAuthenticationSchemeProtocol>
@@ -37,6 +39,14 @@ NS_ASSUME_NONNULL_BEGIN
                         requestUrl:(NSURL *)requestUrl
                              nonce:(nullable NSString *)nonce
               additionalParameters:(nullable NSDictionary *)additionalParameters;
+
+- (instancetype)initWithHttpMethod:(MSALHttpMethod)httpMethod
+                        requestUrl:(NSURL *)requestUrl
+                             nonce:(nullable NSString *)nonce
+              additionalParameters:(nullable NSDictionary *)additionalParameters
+                   externalKeyPair:(nullable MSALExternalKeyPair *)externalKeyPair;
+
+@property (nonatomic, readonly, nullable) MSALExternalKeyPair *externalKeyPair;
 
 - (instancetype)init NS_UNAVAILABLE;
 + (instancetype)new NS_UNAVAILABLE;